Move Index and Date classes to the commons.core.datatypes package
