+++ src/test/java/wanted/commons/core/datatypes/DateTest.java
@@ -1,4 +1,4 @@
-package wanted.commons.core.date;
+package wanted.commons.core.datatypes;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
