+++ src/test/java/wanted/logic/commands/EditCommandTest.java
@@ -17,7 +17,7 @@ import static wanted.testutil.TypicalPersons.getTypicalAddressBook;
 
 import org.junit.jupiter.api.Test;
 
-import wanted.commons.core.index.Index;
+import wanted.commons.core.datatypes.Index;
 import wanted.logic.Messages;
 import wanted.logic.commands.EditCommand.EditPersonDescriptor;
 import wanted.model.LoanBook;
