+++ src/test/java/wanted/testutil/TestUtil.java
@@ -5,7 +5,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
-import wanted.commons.core.index.Index;
+import wanted.commons.core.datatypes.Index;
 import wanted.model.Model;
 import wanted.model.loan.Loan;
